feat: add comprehensive model distillation tutorial with streaming datasets on TPUs
